refact: improve people import
